+++ src-docs/src/views/button/button_example.js
@@ -151,7 +151,7 @@ const buttonGroupSnippet = [
   options={[
     {
       id,
-      label'
+      label
     }
   ]}
 />`,
@@ -163,7 +163,7 @@ const buttonGroupSnippet = [
   options={[
     {
       id,
-      label'
+      label
     }
   ]}
 />`,